# tickets/public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ticket Sales API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000"
        }
    ],
    "paths": {
...
        "/api/purchases": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Get all purchases for authenticated user",
                "description": "Returns all purchases for the logged-in user",
                "operationId": "3e7fda657eab69d8a8f41f202e44a344",
                "responses": {
                    "200": {
                        "description": "List of user purchases"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                }
            },
            "post": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Create purchase with multiple ticket types",
                "description": "Creates purchases for multiple ticket types in a single transaction",
                "operationId": "c5d930515642418b262ea062726e90cf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "event_id",
                                    "tickets"
                                ],
                                "properties": {
                                    "event_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tickets": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "ticket_type_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "quantity": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "example": [
                                            {
                                                "ticket_type_id": 1,
                                                "quantity": 2
                                            },
                                            {
                                                "ticket_type_id": 2,
                                                "quantity": 1
                                            }
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchases created successfully"
                    },
                    "400": {
                        "description": "Not enough tickets or validation error"
                    }
                }
            }
        },
        "/api/purchases/{purchase}": {
            "get": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Get single purchase",
                "description": "Returns purchase details for authenticated user",
                "operationId": "6df95d81052591218d0127db49944326",
                "parameters": [
                    {
                        "name": "purchase",
                        "in": "path",
                        "description": "Purchase ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase details"
                    },
                    "403": {
                        "description": "Unauthorized - not your purchase"
                    },
                    "404": {
                        "description": "Purchase not found"
                    }
                }
            }
        },
        "/api/events/{event}/purchases/reserve": {
            "post": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Reserve tickets for an event",
                "description": "Creates a pending purchase reservation",
                "operationId": "f2721cd96ea47710e56a5431a3295d2f",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_type_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "ticket_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reservation created"
                    },
                    "400": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/purchases/{purchase}/pay": {
            "post": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Pay for a purchase",
                "description": "Marks a pending purchase as completed",
                "operationId": "563085dc972400de3cec4f8fafee7c50",
                "parameters": [
                    {
                        "name": "purchase",
                        "in": "path",
                        "description": "Purchase ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Payment successful"
                    },
                    "400": {
                        "description": "Cannot pay for this purchase"
                    }
                }
            }
        },
        "/api/purchases/{purchase}/cancel": {
            "post": {
                "tags": [
                    "Purchases"
                ],
                "summary": "Cancel a purchase",
                "description": "Cancels a pending purchase",
                "operationId": "e99e1eb389546a897f64282c3a4bfc12",
                "parameters": [
                    {
                        "name": "purchase",
                        "in": "path",
                        "description": "Purchase ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Purchase cancelled"
                    },
                    "400": {
                        "description": "Cannot cancel a completed purchase"
                    }
                }
            }
        },
        "/api/events/{event}/queue/join": {
            "put": {
                "tags": [
                    "Queue"
                ],
                "summary": "Join queue for an event",
                "description": "Adds user to waiting queue (legacy endpoint)",
                "operationId": "8fc97858273b5134beebd428f59ff6a0",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "ticket_type_id"
                                ],
                                "properties": {
                                    "ticket_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Joined queue successfully"
                    }
                }
            }
        },
        "/api/events/{event}/queue/status": {
            "get": {
                "tags": [
                    "Queue"
                ],
                "summary": "Get queue status",
                "description": "Returns user's position in queue (legacy endpoint)",
                "operationId": "54b14adcf5648ac1f538cb5a62e5fc17",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Queue status"
                    }
                }
            }
        },
        "/api/events/{event}/queue/admit": {
            "post": {
                "tags": [
                    "Queue"
                ],
                "summary": "Admit users from queue",
                "description": "Admin only - Admits next users from waiting queue",
                "operationId": "4adb883561dca2aeddd0ae2cd59f2ee2",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "count": {
                                        "description": "Number of users to admit (default 50, max 2000)",
                                        "type": "integer",
                                        "example": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Users admitted successfully"
                    },
                    "403": {
                        "description": "Admin access required"
                    }
                }
            }
        },
...
    },
...
}